pyproject.toml
[tool.poetry]
name = "terminalchat"
version = "0.1.0"
description = "ChatGPT com várias features no terminal"
authors = ["Rinaldo <rinaldomm@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.0"
decouple = "^0.0.7"
openai = "^1.3.2"
tinydb = "^4.8.0"
python-dotenv = "^1.0.0"
bpython = "^0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
